+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+Fix: - Take back input refs on `TextInput`.
+
 ## [2.17.0] - 2019-07-24
 
 Features: