+++ CHANGELOG.md
@@ -4,7 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Update `InfoLines` with styled-components.
+Features:
+- Update `InfoLines` with styled-components.
+- Add `withoutBorder` prop to `RewardCard`.
+- Clicking on `TextCopy input just select the text (no copy in clipboard)
 
 Fix:
 - Take back input refs on `TextInput`.