+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 Features:
 - Update `InfoLines` with styled-components.
 - Add `withoutBorder` prop to `RewardCard`.
-- Clicking on `TextCopy` input just select the text (no copy in clipboard)
+- Clicking on `TextCopy` input just select the text (no copy in clipboard).
 
 Fix:
 - Take back input refs on `TextInput`.