+++ CHANGELOG.md
@@ -4,10 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [2.18.0] - 2019-08-02
+
 Features:
 - Update `InfoLines` with styled-components.
 - Add `withoutBorder` prop to `RewardCard`.
-- Clicking on `TextCopy input just select the text (no copy in clipboard)
+- Clicking on `TextCopy` input just select the text (no copy in clipboard).
 
 Fix:
 - Take back input refs on `TextInput`.